# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Extrittio API",
    "description": "IoT Hub platform REST API",
    "contact": {
      "name": "Extrittio contributors"
    },
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/rules": {
      "get": {
        "tags": [
          "rules"
        ],
        "operationId": "list_rules",
        "parameters": [
          {
            "name": "enabled",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "trigger_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "rules"
        ],
        "operationId": "create_rule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/rules/{id}": {
      "get": {
        "tags": [
          "rules"
        ],
        "operationId": "get_rule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "put": {
        "tags": [
          "rules"
        ],
        "operationId": "update_rule_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "rules"
        ],
        "operationId": "delete_rule_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/rules/{id}/enabled": {
      "put": {
        "tags": [
          "rules"
        ],
        "operationId": "toggle_rule",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnabledInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleResponse"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionInput": {
        "type": "object",
        "required": [
          "action_type",
          "config"
        ],
        "properties": {
          "action_type": {
            "type": "string"
          },
          "config": {}
        }
      },
      "ActionResponse": {
        "type": "object",
        "required": [
          "id",
          "action_type",
          "config"
        ],
        "properties": {
          "action_type": {
            "type": "string"
          },
          "config": {},
          "id": {
            "type": "string"
          }
        }
      },
...
      "ConditionInput": {
        "type": "object",
        "required": [
          "field",
          "operator",
          "value"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "operator": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
      "ConditionResponse": {
        "type": "object",
        "required": [
          "id",
          "field",
          "operator",
          "value"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "operator": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "CreateRuleRequest": {
        "type": "object",
        "required": [
          "name",
          "trigger_type",
          "target_type",
          "conditions",
          "actions"
        ],
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionInput"
            }
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConditionInput"
            }
          },
          "cooldown_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "target_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "target_type": {
            "type": "string"
          },
          "trigger_type": {
            "type": "string"
          }
        }
      },
...
      "EnabledInput": {
        "type": "object",
        "required": [
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
      "RuleResponse": {
        "type": "object",
        "required": [
          "id",
          "name",
          "enabled",
          "trigger_type",
          "target_type",
          "cooldown_seconds",
          "conditions",
          "actions",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionResponse"
            }
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConditionResponse"
            }
          },
          "cooldown_seconds": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "target_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "target_type": {
            "type": "string"
          },
          "trigger_type": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "UpdateRuleRequest": {
        "type": "object",
        "properties": {
          "actions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/ActionInput"
            }
          },
          "conditions": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/ConditionInput"
            }
          },
          "cooldown_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "target_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "target_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "trigger_type": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT bearer authentication for explicitly enabled non-browser clients"
      },
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "extrittio_session",
        "description": "HTTP-only browser session cookie issued by POST /api/v1/auth/login"
      }
    }
  },
...
}